backend/requirements.txt
# --- Framework Web & Servidor ---
fastapi>=0.115.0        # Versão mais atual para compatibilidade 3.13 Framework moderno para construção de APIs rápidas
uvicorn[standard]>=0.30.0       # Servidor ASGI de alta performance para rodar o FastAPI

# --- Processamento de Dados & ETL ---
pandas>=2.2.2          # Manipulação e análise de dados (essencial para os CSVs da ANS)
requests>=2.32.0        # Realização de requisições HTTP (para baixar os ZIPs da API)
openpyxl>=3.1.5        # Suporte para leitura/escrita de arquivos Excel (.xlsx)

# --- Banco de Dados & ORM ---
sqlalchemy>=2.0.30     # Toolkit SQL e ORM para mapeamento de tabelas
psycopg2-binary>=2.9.9  # Driver de conexão para bancos de dados PostgreSQL

# --- Validação & Configurações ---
pydantic>=2.9.0         # Validação de dados e definição de schemas (Data Integrity)
pydantic-settings>=2.5.0 # Gestão de variáveis de ambiente integrada ao Pydantic
python-dotenv>=1.0.1    # Carregamento de variáveis de ambiente a partir do arquivo .env
beautifulsoup4>=4.12.0 # Extração de links e raspagem de dados